requirements.txt
appdirs==1.4.4
apscheduler==2.1.2
arabic-reshaper==2.1.3
attrs==21.4.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
bottle==0.11.6
cached-property==1.5.2
cachetools==3.1.1
certifi==2021.10.8
chardet==4.0.0; python_version < '3'
clickhouse-driver==0.2.0
configparser==4.0.2; python_version < '3'
crypto==1.4.1
defusedxml==0.7.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
enum34==1.1.10; python_version < '3.4'
future==0.18.2; python_version >= '2.6' and python_version not in '3.0, 3.1, 3.2, 3.3'
futures==3.3.0; python_version < '3.2'
gevent==1.3b1
google-api-core==1.31.6; python_version < '3'
google-api-python-client==1.12.11
google-auth-httplib2==0.1.0
google-auth==1.35.0; python_version < '3'
googleapis-common-protos==1.52.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
greenlet==1.1.2; platform_python_implementation == 'CPython'
gspread==0.2.5
html5lib==1.1
httplib2==0.20.4; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
https://pypi.python.org/packages/source/o/oauth2client/oauth2client-2.0.0.post1.tar.gz#egg=oauth2client
idna==2.10; python_version < '3'
ipaddress==1.0.23; python_version < '3.4'
isodate==0.6.1
lxml==4.9.0
naked==0.1.31
packaging==20.9; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
pillow==6.2.2; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
protobuf==3.17.3; python_version < '3'
psutil==5.9.1
pyasn1-modules==0.2.8
pyasn1==0.4.8
pycrypto==2.6.1
pymongo==3.2.2
pyparsing==2.4.7; python_version < '3'
pypdf2==1.28.4; python_version >= '2.7'
python-bidi==0.4.2
python-dateutil==2.2
pytz==2013.8
pyyaml==5.4.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5'
redis==2.8.0
reportlab==3.5.59
requests-toolbelt==0.9.1
requests==2.27.1
rq==0.3.11
rsa==4.5
shellescape==3.8.1
shortuuid==0.4
six==1.13.0
times==0.6.2
typing==3.10.0.0
tzlocal==2.1
uritemplate==3.0.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
urllib3==1.26.9
webencodings==0.5.1
xhtml2pdf==0.2.5
xlrd==2.0.1
xlutils==2.0.0
xlwt==1.3.0
zeep==3.4.0
